+++ ca/django_ca/subject.py
@@ -19,10 +19,7 @@ from typing import Any, Dict, Iterable, Iterator, List, Optional, Tuple, Union, 
 
 from cryptography import x509
 
-from django.core.exceptions import ImproperlyConfigured
-
-from . import ca_settings
-from .deprecation import RemovedInDjangoCA123Warning, RemovedInDjangoCA124Warning
+from .deprecation import RemovedInDjangoCA124Warning
 from .typehints import ParsableSubject
 from .utils import MULTIPLE_OIDS, NAME_OID_MAPPINGS, OID_NAME_MAPPINGS, SUBJECT_FIELDS, parse_name_x509
 
@@ -281,22 +278,3 @@ class Subject:
         <Name(C=AT,CN=example.com)>
         """
         return x509.Name([x509.NameAttribute(k, v) for k, v in self.fields])
-
-
-def get_default_subject() -> Subject:  # pragma: no cover
-    """Get the default subject as configured by the ``CA_DEFAULT_SUBJECT`` setting.
-
-    .. deprecated:: 1.22.0
-
-       This function will be removed in 1.23.0.
-    """
-
-    warnings.warn(
-        "django_ca.subject.get_default_subject() will be removed in 1.23.0.",
-        category=RemovedInDjangoCA123Warning,
-    )
-
-    try:
-        return Subject(ca_settings.CA_DEFAULT_SUBJECT)
-    except (ValueError, KeyError) as ex:
-        raise ImproperlyConfigured(f"CA_DEFAULT_SUBJECT: {ex}") from ex